+++ app/models/item.rb
@@ -3,4 +3,18 @@ class Item < ApplicationRecord
   validates :name, presence: true, length: { maximum: 255 }
   validates :url, presence: true, length: { maximum: 255 }
   validates :image_url, presence: true, length: { maximum: 255 }
-end
+  
+    def read(result)
+      code = result['itemCode']
+      name = result['itemName']
+      url = result['itemUrl']
+      image_url = result['mediumImageUrls'].first.gsub('?_ex=128x128', '')
+
+      {
+        code: code,
+        name: name,
+        url: url,
+        image_url: image_url,
+      }
+    end
+  end
