+++ app/views/layouts/_navbar.html.erb
@@ -9,7 +9,7 @@
         <% if logged_in? %>
           <li>
           　<i class="fas fa-arrow-alt-circle-down"></i>
-            <%= link_to "アイテムを追加", new_item_path %>　　　　　　　
+            <%= link_to "アイテムを追加", items_path %>　　　　　　　
           </li>
           <li>
           　<i class="fas fa-users"></i>
