+++ config/routes.rb
@@ -5,5 +5,5 @@ Rails.application.routes.draw do
   delete 'logout', to: 'sessions#destroy'
   get 'signup', to: 'users#new'
   resources :users, only: [:show, :create, :edit, :update] 
-  resources :items, only: [:new]
+  resources :items, only: [:index]
 end
